Rename constant files
