+++ website/src/constants/index.js
@@ -1,4 +1,4 @@
-export * from "./codeConstants"
-export * from "./featuresConst"
-export * from "./propsConst"
-export * from "./footerConstants"
+export * from './codeConstants';
+export * from './featuresConstants';
+export * from './propsConstants';
+export * from './footerConstants';
